app/requirements.txt
# app/requirements.txt
flask
psycopg2-binary
flask-sqlalchemy
python-dotenv
